+++ src/webpack/static-render-pages.js
@@ -27,18 +27,7 @@ function staticRenderPages(
   cssUrl?: string
 ): Promise<Array<void>> {
   const inlineJsScripts = renderInlineJsScripts(batfishConfig.inlineJs);
-
-  // Load the full stylesheet lazily, after DOMContentLoaded. The page will
-  // still render quickly because it will have its own CSS injected inline.
-  let css = '';
-  if (cssUrl) {
-    if (batfishConfig.staticHtmlInlineDeferCss) {
-      const loadCssJs = `document.addEventListener('DOMContentLoaded',function(){var s=document.createElement('link');s.rel='stylesheet';s.href='${cssUrl}';document.head.insertBefore(s, document.getElementById('loadCss'));});`;
-      css = `<script id="loadCss">${loadCssJs}</script>`;
-    } else {
-      css = `<link rel="stylesheet" href="${cssUrl}" />`;
-    }
-  }
+  const css = cssUrl ? `<link rel="stylesheet" href="${cssUrl}" />` : '';
 
   const appendToBody = [
     // The Webpack manifest is inlined because it is very small.